+++ X10D.Tests/src/Text/CharTests.cs
@@ -1,4 +1,5 @@
-﻿using Microsoft.VisualStudio.TestTools.UnitTesting;
+﻿using System.Text;
+using Microsoft.VisualStudio.TestTools.UnitTesting;
 using X10D.Text;
 
 namespace X10D.Tests.Text;
@@ -6,6 +7,25 @@ namespace X10D.Tests.Text;
 [TestClass]
 public class CharTests
 {
+    [TestMethod]
+    public void IsEmoji_ShouldReturnTrue_GivenBasicEmoji()
+    {
+        Assert.IsTrue('✂'.IsEmoji());
+        Assert.IsTrue('✅'.IsEmoji());
+        Assert.IsTrue('❎'.IsEmoji());
+        Assert.IsTrue('➕'.IsEmoji());
+        Assert.IsTrue('➖'.IsEmoji());
+    }
+
+    [TestMethod]
+    public void IsEmoji_ShouldReturnFalse_GivenNonEmoji()
+    {
+        for (var letter = 'A'; letter <= 'Z'; letter++)
+        {
+            Assert.IsFalse(letter.IsEmoji());
+        }
+    }
+
     [TestMethod]
     public void RepeatShouldBeCorrect()
     {